+++ src/global/Vector100Dtype.java
@@ -2,21 +2,21 @@ package global;
 
 public class Vector100Dtype {
 
-  private final static short MIN_VAL = -10000;
-  private final static short MAX_VAL = 10000;
-  private short[] dimension;
+  private final static int MIN_VAL = -10000;
+  private final static int MAX_VAL = 10000;
+  private int[] dimension;
 
   public Vector100Dtype() {
-    dimension = new short[100];
+    dimension = new int[100];
   }
 
-  public Vector100Dtype(short[] dimension) {
+  public Vector100Dtype(int[] dimension) {
     this.dimension = dimension;
   }
-  public short[] getDimension() {
+  public int[] getDimension() {
     return dimension;
   }
-  public void setDimension(short[] dimension) {
+  public void setDimension(int[] dimension) {
     this.dimension = dimension;
   }
 }
